+++ test/e2e/raja/e2e.md
@@ -1,5 +1,5 @@
-```mermaid 
+```mermaid
 sequenceDiagram
 	actor User
-	User ->> main : aws-env go run main.go start 
+	User ->> main : aws-env go run main.go start
 ```
